add ability to change wool of the boots change how things register piglins now ignore gold armor

+++ common/src/main/java/com/msg/combining_gears/items/ElytraChestplate.java
@@ -22,7 +22,7 @@ import net.minecraft.world.level.gameevent.GameEvent;
 
 public class ElytraChestplate extends ArmorItem {
 
-    public ElytraChestplate(String name, int defense, int enchantmentValue, Holder<SoundEvent> sound, ItemLike repairItem, int toughness, int knockbackResistance, Properties properties) {
+    public ElytraChestplate(String name, int defense, int enchantmentValue, Holder<SoundEvent> sound, ItemLike repairItem, int toughness, float knockbackResistance, Properties properties) {
         super(Registry.registerForHolder(BuiltInRegistries.ARMOR_MATERIAL,
                                                                         Common.resourcesLocation(name),
                                                                         new ArmorMaterial(
